+++ packages/fabric/emoji/src/components/picker/EmojiPickerFooter.tsx
@@ -22,6 +22,7 @@ export interface Props {
   uploadErrorMessage?: string;
   onUploadCancelled: () => void;
   onUploadEmoji: OnUploadEmoji;
+  onFileChosen?: (name: string) => void;
 }
 
 export default class EmojiPickerFooter extends PureComponent<Props, {}> {
@@ -36,6 +37,7 @@ export default class EmojiPickerFooter extends PureComponent<Props, {}> {
       toneEmoji,
       uploadErrorMessage,
       uploading,
+      onFileChosen,
     } = this.props;
 
     const previewFooterClassnames = classNames([
@@ -49,6 +51,7 @@ export default class EmojiPickerFooter extends PureComponent<Props, {}> {
           <EmojiUploadPicker
             onUploadCancelled={onUploadCancelled}
             onUploadEmoji={onUploadEmoji}
+            onFileChosen={onFileChosen}
             errorMessage={uploadErrorMessage}
             initialUploadName={initialUploadName}
           />