+++ packages/fabric/emoji/src/constants.ts
@@ -30,3 +30,5 @@ export const defaultCategories = [
 export const defaultListLimit = 50;
 
 export const migrationUserId = 'hipchat_migration_emoticons';
+
+export const analyticsEmojiPrefix = 'atlassian.fabric.emoji.picker';