+++ packages/fabric/mention/docs/0-intro.tsx
@@ -61,7 +61,6 @@ export default md`
 
   * Promise ([polyfill](https://www.npmjs.com/package/es6-promise), [browser support](http://caniuse.com/#feat=promises))
   * Fetch API ([polyfill](https://www.npmjs.com/package/whatwg-fetch), [browser support](http://caniuse.com/#feat=fetch))
-  * URLSearchParams API ([polyfill](https://www.npmjs.com/package/url-search-params), [browser support](http://caniuse.com/#feat=urlsearchparams))
 
   If a \`target\` property is provided with a \`position\` property, then the
   Picker will automatically be positioned floating above that element. The \`target\`