+++ packages/fabric/conversation/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @atlaskit/conversation
 
+## 5.1.3
+
 ## 5.1.2
 
 ## 5.1.1